requirements.txt
Flask==3.0.0
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
slack-bolt==1.18.1
slack-sdk==3.26.0
urllib3==2.1.0
gunicorn
